+++ packages/php-wasm/node/src/lib/index.ts
@@ -1,4 +1,4 @@
 export * from './get-php-loader-module';
 export * from './parse-worker-startup-options';
 export * from './networking/with-networking';
-export * from './php';
+export * from './node-php';
